feat: use azure models for suggestion and dicovery
