+++ lib/workflow/models.ts
@@ -1,0 +1,13 @@
+import { createOpenRouter } from '@openrouter/ai-sdk-provider';
+
+export const modelProviders = () => {
+  const openRouter = createOpenRouter({
+    apiKey: process.env.OPENROUTER_API_KEY,
+  });
+
+  return {
+    discovery: openRouter('openai/gpt-4.1-mini'),
+    workflow: openRouter('google/gemini-2.5-pro-preview'),
+    suggestion: openRouter('openai/gpt-4.1'),
+  };
+};
